Fix types

+++ pkg/models/projects/types.go
@@ -51,7 +51,7 @@ type Project struct {
 	// TestUserLockTTL is an int representing the time in seconds that the user in test remains locked once the lock is set. Defaults to 1hr (3600sec)
 	TestUserLockTTL int64 `json:"test_user_lock_ttl"`
 	// LiveUserLockTTL is an int representing the time in seconds that the user in live remains locked once the lock is set. Defaults to 1hr (3600sec)
-	LiveUserLockTTL int32 `json:"live_user_lock_ttl"`
+	LiveUserLockTTL int64 `json:"live_user_lock_ttl"`
 }
 
 type CreateRequest struct {
@@ -78,7 +78,7 @@ type CreateRequest struct {
 	// TestUserLockTTL is an int representing the time in seconds that the user in test remains locked once the lock is set. Defaults to 1hr (3600sec) if omitted.
 	TestUserLockTTL *int64 `json:"test_user_lock_ttl"`
 	// LiveUserLockTTL is an int representing the time in seconds that the user in live remains locked once the lock is set. Defaults to 1hr (3600sec) if omitted.
-	LiveUserLockTTL *int32 `json:"live_user_lock_ttl"`
+	LiveUserLockTTL *int64 `json:"live_user_lock_ttl"`
 }
 
 type CreateResponse struct {
@@ -151,7 +151,7 @@ type UpdateRequest struct {
 	// TestUserLockTTL is an int representing the time in seconds that the user in test remains locked once the lock is set. Defaults to 1hr (3600sec)
 	TestUserLockTTL *int64 `json:"test_user_lock_ttl"`
 	// LiveUserLockTTL is an int representing the time in seconds that the user in live remains locked once the lock is set. Defaults to 1hr (3600sec)
-	LiveUserLockTTL *int32 `json:"live_user_lock_ttl"`
+	LiveUserLockTTL *int64 `json:"live_user_lock_ttl"`
 }
 
 type UpdateResponse struct {
